feat: add jwt and role based authentication to endpoint

+++ src/users/users.controller.ts
@@ -1,8 +1,20 @@
-import { Controller, Delete, Get, Param } from '@nestjs/common';
+import {
+  Controller,
+  Delete,
+  Get,
+  HttpException,
+  Param,
+  Req,
+  UseGuards,
+} from '@nestjs/common';
 import { UserService } from './users.service';
 import { Post, Body, HttpCode, HttpStatus } from '@nestjs/common';
 import type { CreateUserDTO } from './dto/create-user.dto';
 import GetUserDTO from './dto/get-user.dto';
+import type { Request } from 'express';
+import { JwtAuthGuard } from 'src/guards/jwt-auth.guard';
+import { RolesGuard } from 'src/guards/roles.guard';
+import { Roles } from 'src/decorators/roles.decorator';
 
 @Controller('users')
 export class UserController {
@@ -12,26 +24,37 @@ export class UserController {
   @HttpCode(HttpStatus.CREATED)
   async create(@Body() createUserDto: CreateUserDTO): Promise<object> {
     await this.userService.create(createUserDto);
-    return {statusCode: 201, message: 'User created succesfully' };
+    return { statusCode: 201, message: 'User created succesfully' };
   }
 
   @Get()
   @HttpCode(HttpStatus.OK)
-  async getUsers() : Promise<Array<GetUserDTO>> {
+  async getUsers(): Promise<Array<GetUserDTO>> {
     return await this.userService.getUsers();
   }
 
   @Get(':id')
   @HttpCode(HttpStatus.OK)
-  async getUserById(@Param('id') userId : string) : Promise<GetUserDTO> {
+  async getUserById(@Param('id') userId: string): Promise<GetUserDTO> {
     return await this.userService.getUserById(userId);
   }
 
   @Delete(':id')
   @HttpCode(HttpStatus.ACCEPTED)
-  async deleteUserById(@Param('id') userId : string): Promise<object> {
+  @Roles(['admin'])
+  @UseGuards(JwtAuthGuard, RolesGuard)
+  async deleteUserById(
+    @Param('id') userId: string,
+    @Req() req: Request,
+  ): Promise<object> {
+    const currentUser = req.user;
+    if (req.user.userId === userId) {
+      throw new HttpException(
+        'You cannot delete your own user',
+        HttpStatus.FORBIDDEN,
+      );
+    }
     await this.userService.deleteUserById(userId);
-    return {statusCode: 202, message : "User deleted succesfully"}
+    return { statusCode: 202, message: 'User deleted succesfully' };
   }
-  
 }
